% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{calculate_percentage}
\alias{calculate_percentage}
\title{Calculate a percentage}
\usage{
calculate_percentage(x, y, d)
}
\arguments{
\item{x}{A number}

\item{y}{A number}

\item{d}{An integer}
}
\value{
The ratio of x/y expressed as a percentage with d digits
}
\description{
\code{calculate_percentage} calculates a ratio and express it as a percentage
}
\examples{
calculate_percentage(27, 80, 1)

}
